Minor refactor of healthcheck component
