+++ frontend/src/App.tsx
@@ -13,7 +13,7 @@ import { BackofficeMode } from './api/BackofficeMode'
 import NamespaceContext from './domain/context/NamespaceContext'
 import { ErrorToastNotification } from './features/commonComponents/ErrorToastNotification'
 import FavoriteVessels from './features/favorite_vessels/FavoriteVessels'
-import { Healthcheck } from './features/healthcheck/Healthcheck'
+import { Healthcheck } from './features/Healthcheck'
 import LayersSidebar from './features/layers/LayersSidebar'
 import { DrawLayerModal } from './features/map/draw/DrawModal'
 import Map from './features/map/Map'
